refactor: Dashboard 상태 관련 로직 분리
Dashboard.tsx에 있던 상태관련 로직을 useDashboard.tsx파일로 옮겼습니다
