+++ src/components/Dashboard/Dashboard.tsx
@@ -1,16 +1,9 @@
-import { useHabitStateContext } from 'contexts/HabitContext';
-import { getDayword } from 'lib/helpers/date';
-import { getRoutineByDate } from 'lib/helpers/habit';
 import styles from './Dashboard.module.css';
 import HabitCard from './HabitCard/HabitCard';
+import useDashboard from './useDashboard';
 
 const DashBoard = () => {
-  const habits = useHabitStateContext();
-  const todayHabits = habits.filter((habit) => {
-    const todayDate = new Date();
-    const routine = getRoutineByDate(habit.routineList, todayDate);
-    return routine?.includes(getDayword(todayDate)) ? true : false;
-  });
+  const todayHabits = useDashboard();
 
   return (
     <div className={styles.container}>
